+++ src/features/dashboard/JobListClient.tsx
@@ -6,6 +6,7 @@ import { useJobFitFilter } from './hooks/useJobFitFilter';
 import { useBookmarkToggle } from './hooks/useBookmarkToggle';
 import { JobListSection } from './ui/JobListSection';
 import { ConfirmModal } from '@/shared/ui/ConfirmModal';
+import { ROUTES } from '@/shared/constants/routes';
 
 type JobListClientProps = {
   profileProvinces: string[];
@@ -56,7 +57,7 @@ export function JobListClient({
           confirmLabel="로그인하기"
           cancelLabel="닫기"
           onConfirm={() => {
-            window.location.href = '/api/oauth/kakao/authorize';
+            window.location.href = ROUTES.KAKAO_AUTHORIZE;
           }}
           onClose={closeLoginModal}
         />
