+++ src/shared/constants/routes.ts
@@ -1,0 +1,3 @@
+export const ROUTES = {
+  KAKAO_AUTHORIZE: '/api/oauth/kakao/authorize',
+} as const;
